requirements.txt
Flask==3.0.0
firebase-admin==6.2.0
google-cloud-secret-manager>=2.16.0
google-generativeai==0.8.3
google-cloud-firestore==2.15.0
functions-framework==3.*